+++ src/VexLibrary/utils/pose.cpp
@@ -2,7 +2,7 @@
 
 void Lib::Pose::set_x(float x) { this->x = x; };
 void Lib::Pose::set_y(float y) { this->y = y; };
-void Lib::Pose::set_theta(float theta) { this->theta = theta; };
+void Lib::Pose::set_theta(double theta) { this->theta = theta; };
 void Lib::Pose::set_pose(settings s) {
   this->x = s.x;
   this->y = s.y;
@@ -11,4 +11,4 @@ void Lib::Pose::set_pose(settings s) {
 
 float Lib::Pose::get_x() const { return this->x; };
 float Lib::Pose::get_y() const { return this->y; };
-float Lib::Pose::get_theta() const { return this->theta; };
+double Lib::Pose::get_theta() const { return this->theta; };
